transcribe_app/requirements.txt
openai-whisper==20231117
fastapi==0.111.0